# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi Main API",
        "description": "Tulang punggung dari berbagai aplikasi.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/motorinci/reviews": {
            "get": {
                "tags": [
                    "Motorinci Reviews"
                ],
                "summary": "Mendapatkan daftar ulasan motorinci",
                "operationId": "getMotorinciReviewsList",
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200_Success"
                    },
                    "401": {
                        "$ref": "#/components/responses/401_Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Motorinci Reviews"
                ],
                "summary": "Membuat ulasan motorinci baru",
                "operationId": "storeMotorinciReview",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "motor_id",
                                    "reviewer_name",
                                    "reviewer_email",
                                    "rating"
                                ],
                                "properties": {
                                    "motor_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "reviewer_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "reviewer_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "rating": {
                                        "description": "Nilai rating dari 1 sampai 5",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Motor ini sangat bagus dan hemat bahan bakar."
                                    },
                                    "is_approved": {
                                        "description": "Status persetujuan ulasan",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/201_Created"
                    },
                    "422": {
                        "$ref": "#/components/responses/422_UnprocessableContent"
                    },
                    "401": {
                        "$ref": "#/components/responses/401_Unauthorized"
                    }
                }
            }
        },
        "/api/motorinci/reviews/{id}": {
            "get": {
                "tags": [
                    "Motorinci Reviews"
                ],
                "summary": "Mendapatkan satu ulasan motorinci berdasarkan ID",
                "operationId": "getMotorinciReviewById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Motorinci review ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200_Success"
                    },
                    "404": {
                        "$ref": "#/components/responses/404_NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/401_Unauthorized"
                    }
                }
            },
            "put": {
                "tags": [
                    "Motorinci Reviews"
                ],
                "summary": "Memperbarui ulasan motorinci yang sudah ada",
                "operationId": "updateMotorinciReview",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Motorinci review ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "reviewer_name": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "reviewer_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jane.doe@example.com"
                                    },
                                    "rating": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Cukup baik, tapi kurang responsif."
                                    },
                                    "is_approved": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200_Success"
                    },
                    "404": {
                        "$ref": "#/components/responses/404_NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/422_UnprocessableContent"
                    },
                    "401": {
                        "$ref": "#/components/responses/401_Unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Motorinci Reviews"
                ],
                "summary": "Menghapus ulasan motorinci",
                "operationId": "deleteMotorinciReview",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Motorinci review ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "$ref": "#/components/responses/204_NoContent"
                    },
                    "404": {
                        "$ref": "#/components/responses/404_NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/401_Unauthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "SuccessJsonResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Operasi berhasil"
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ErrorJsonResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Terjadi kesalahan"
                    }
                },
                "type": "object"
            },
            "ValidationErrorJsonResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Data yang diberikan tidak valid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "field_name": [
                                "Pesan error."
                            ]
                        }
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "200_Success": {
                "description": "Operasi berhasil dieksekusi.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/SuccessJsonResponse"
                        }
                    }
                }
            },
            "201_Created": {
                "description": "Sumber daya berhasil dibuat.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/SuccessJsonResponse"
                        }
                    }
                }
            },
            "204_NoContent": {
                "description": "Operasi berhasil tetapi tidak ada konten untuk dikembalikan."
            },
            "401_Unauthorized": {
                "description": "Autentikasi gagal. Token tidak valid atau tidak ada.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorJsonResponse"
                        }
                    }
                }
            },
...
            "404_NotFound": {
                "description": "Sumber daya yang diminta tidak ditemukan.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorJsonResponse"
                        }
                    }
                }
            },
            "422_UnprocessableContent": {
                "description": "Validasi gagal.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ValidationErrorJsonResponse"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token Bearer Anda. Contoh: 'Bearer 1|abcdef...'",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}